refactor: disambiguate session-stores naming collision
src/tools/agentCliSessionStores.ts exported no store: it registered host
shutdown handlers plus per-session WeakMap accessors for the CLI-session
registries. Rename it to agentShutdownHandlers.ts to match what it does.

src/controllers/session/sessionStores.ts differed from
src/agent/storage/SessionStores.ts only by case, both independently
plausible as "the session stores module" on a grep. Rename it to
createSessionStores.ts to match its single factory export.

Pure rename plus import-path updates; no behavior change.
