+++ packages/cli/src/runtime/initPlatform.ts
@@ -24,7 +24,7 @@ import {
 import { openTexraConfigStores } from '@platform/defaults/nodeStores';
 import { GlobalStateKey } from '@shared/state/stateKeys';
 import { UsageLogService } from '@telemetry/UsageLogService';
-import { registerAgentShutdownHandlers } from '@tools/agentCliSessionStores';
+import { registerAgentShutdownHandlers } from '@tools/agentShutdownHandlers';
 import { seedDisabledToolDefaults } from '@tools/toolAvailability';
 import { setSetupPlatform } from '@tools/setup/platform';
 import { toErrorMessage } from '@utils/errors/errorMessage';
