+++ src/tools/agentShutdownHandlers.ts
